requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.7
django-import-export==4.1.1
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
